fix: address PR review — color fallback, move CATEGORY_COLORS, translate UI strings
- Add ?? #888888 fallback for CATEGORY_COLORS[code] in category tile
  and subcategory pill to prevent transparent/broken shadow on unknown codes
- Move CATEGORY_COLORS definition to src/lib/content/contentText.tsx
  (alongside categoryOptions/subcategoriesMap); preview/constants.ts
  re-exports it so preview components need no changes
- Update StepEventDetails import to use contentText as source of truth
- Translate Swedish help text to English to match surrounding UI labels
